+++ crates/aiken-lang/src/parser/literal/mod.rs
@@ -1,3 +1,7 @@
-pub(crate) mod bytearray;
-pub(crate) mod int;
-pub(crate) mod string;
+mod bytearray;
+mod string;
+mod uint;
+
+pub use bytearray::{array_of_bytes, hex_string, parser as bytearray, utf8_string};
+pub use string::parser as string;
+pub use uint::parser as uint;
